+++ toys/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.3.8 / TBD
 
+* CHANGED: Renamed helpers to mixins.
 * CHANGED: Renamed :in_from, :out_to, and :err_to exec options to :in, :out, :err
 * IMPROVED: Exec raises an error if passed an unknown option.
 * IMPROVED: Exec now accepts nearly all the same stream specifications as Process#spawn.